+++ src/App.tsx
@@ -11,6 +11,10 @@ import Room from './components/Room/Room';
 import useHeight from './hooks/useHeight/useHeight';
 import useRoomState from './hooks/useRoomState/useRoomState';
 
+import { Logger } from 'twilio-video';
+
+Logger.getLogger('twilio-video').setLevel('WARN');
+
 const Container = styled('div')({
   display: 'grid',
   gridTemplateRows: '1fr auto',